Move out file writing to a separate class.
